refactor(Explorers): removed chain.so & split APIs in separate files
